+++ apps/minifront/src/components/extension-unavailable.tsx
@@ -9,7 +9,7 @@ export const ExtensionUnavailable = () => {
     <>
       <HeadTag />
 
-      <SplashPage title='Penumbra extension unavailble'>
+      <SplashPage title='Penumbra extension unavailable'>
         <p className='mb-2'>We can&apos;t currently connect to the Penumbra extension.</p>
         <p className='mb-2'>
           This page may have been left open for too long, causing a timeout. Please reload this page
