+++ apps/minifront/src/components/ibc/ibc-in-form.tsx
@@ -1,0 +1,9 @@
+import { Card } from '@penumbra-zone/ui';
+
+export const IbcInForm = () => {
+  return (
+    <Card gradient className='md:p-5'>
+      To come...
+    </Card>
+  );
+};
