+++ apps/minifront/src/components/metadata/content.ts
@@ -39,10 +39,6 @@ export const metadata: Record<PagePath, PageMetadata> = {
     title: 'Penumbra | Swap',
     description: eduPanelContent[EduPanel.TEMP_FILLER],
   },
-  [PagePath.GOVERNANCE]: {
-    title: 'Penumbra | Governance',
-    description: eduPanelContent[EduPanel.TEMP_FILLER],
-  },
   [PagePath.STAKING]: {
     title: 'Penumbra | Staking',
     description: eduPanelContent[EduPanel.TEMP_FILLER],
