+++ apps/minifront/src/components/metadata/paths.ts
@@ -2,12 +2,11 @@ export enum PagePath {
   INDEX = '/',
   SWAP = '/swap',
   SEND = '/send',
-  GOVERNANCE = '/governance',
   STAKING = '/staking',
   RECEIVE = '/send/receive',
   TRANSACTIONS = '/dashboard/transactions',
   DASHBOARD = '/dashboard',
-  IBC = '/send/ibc',
+  IBC = '/ibc',
   NFTS = '/dashboard/nfts',
   TRANSACTION_DETAILS = '/tx/:hash',
 }
