+++ apps/minifront/src/components/send/constants.ts
@@ -13,15 +13,9 @@ export const sendTabsHelper: SendTabMap = {
     label: 'Receiving Funds',
     content: EduPanel.RECEIVING_FUNDS,
   },
-  [PagePath.IBC]: {
-    src: './ibc-gradient.svg',
-    label: 'IBC withdraw',
-    content: EduPanel.IBC_WITHDRAW,
-  },
 };
 
 export const sendTabs = [
   { title: 'Send', href: PagePath.SEND, active: true },
   { title: 'Receive', href: PagePath.RECEIVE, active: true },
-  { title: 'IBC', href: PagePath.IBC, active: false },
 ];
