+++ apps/minifront/src/components/send/types.ts
@@ -1,7 +1,7 @@
 import { PagePath } from '../metadata/paths';
 import { EduPanel } from '../shared/edu-panels/content';
 
-export type SendTab = PagePath.SEND | PagePath.RECEIVE | PagePath.IBC;
+export type SendTab = PagePath.SEND | PagePath.RECEIVE;
 
 interface SendTabMetadata {
   src: string;
